refactor: generalize imports
